requirements.txt
pymongo[srv]==4.2.0
certifi
python-dotenv